Update listenOnSomeOne.py
Added option 4 for both MITM and spoofing to work at the same time to make the program detect ip addresses that fails in a normal nmap scan. (This Option as the same as option 1 but in this case the code  will cause a denial of service due to the amount of traffic targeted a victim).

Fixed  the  interaction with the menu after the termination of each program.